chore: renamed bgscore to core
