+++ src/main/java/me/remag501/adventure/listener/ExtractionListener.java
@@ -7,8 +7,8 @@ import me.remag501.adventure.model.ExtractionZone;
 import me.remag501.adventure.model.RotationTrack;
 import me.remag501.adventure.setting.SettingsProvider;
 import me.remag501.adventure.util.MessageUtil;
-import me.remag501.bgscore.api.event.EventService;
-import me.remag501.bgscore.api.task.TaskService;
+import me.remag501.core.api.event.EventService;
+import me.remag501.core.api.task.TaskService;
 import org.bukkit.*;
 import org.bukkit.boss.BarColor;
 import org.bukkit.boss.BarStyle;
