+++ src/main/java/me/remag501/adventure/listener/GuiListener.java
@@ -5,8 +5,8 @@ import me.remag501.adventure.manager.RotationManager;
 import me.remag501.adventure.model.RotationTrack;
 import me.remag501.adventure.setting.SettingsProvider;
 import me.remag501.adventure.util.MessageUtil;
-import me.remag501.bgscore.api.event.EventService;
-import me.remag501.bgscore.api.namespace.NamespaceService;
+import me.remag501.core.api.event.EventService;
+import me.remag501.core.api.namespace.NamespaceService;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
