+++ src/main/java/me/remag501/adventure/listener/JoinListener.java
@@ -3,7 +3,7 @@ package me.remag501.adventure.listener;
 import me.remag501.adventure.manager.PDCManager;
 import me.remag501.adventure.manager.PenaltyManager;
 import me.remag501.adventure.manager.RotationManager;
-import me.remag501.bgscore.api.event.EventService;
+import me.remag501.core.api.event.EventService;
 import org.bukkit.Bukkit;
 import org.bukkit.World;
 import org.bukkit.entity.Player;
