+++ src/main/java/me/remag501/adventure/manager/PenaltyManager.java
@@ -5,7 +5,7 @@ import me.remag501.adventure.setting.AdventureSettings;
 import me.remag501.adventure.setting.SettingsProvider;
 import me.remag501.adventure.task.BroadcastTask;
 import me.remag501.adventure.util.MessageUtil;
-import me.remag501.bgscore.api.task.TaskService;
+import me.remag501.core.api.task.TaskService;
 import org.bukkit.*;
 import org.bukkit.entity.Player;
 import org.bukkit.potion.PotionEffect;
