+++ src/main/java/me/remag501/adventure/manager/RotationManager.java
@@ -5,7 +5,7 @@ import me.remag501.adventure.setting.AdventureSettings;
 import me.remag501.adventure.model.RotationTrack;
 import me.remag501.adventure.model.WorldInfo;
 import me.remag501.adventure.setting.SettingsProvider;
-import me.remag501.bgscore.api.task.TaskService;
+import me.remag501.core.api.task.TaskService;
 import org.bukkit.World;
 
 import java.util.*;
